twins-content-engine/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "twins-content-engine"
version = "0.1.0"
description = "Content strategy + generation engine for Twins Garage Doors"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["engine"]

[tool.hatch.build.targets.wheel.force-include]
"engine/schema.sql" = "engine/schema.sql"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
markers = [
    "integration: hits real Anthropic API (gated by INTEGRATION=1)",
]
